OMAP3: PM: Implement DVFS functions used by DSPbridge to control OPP
These functions are called by the DSPbridge to determine the available OPPs,
the current OPP and to request a OPP based on DSP activity.  When a
frequency is requested, the higher of the MPU and DSP requested OPP is used
to pick the used OPP.  A call by the function used by the governor is also
added so that the requested MPU OPP may be known.

Change-Id: I1d9e4c6e7d033c68979579627f51f364391da048
Signed-off-by: Steve Schefter <sschefter@ti.com>