ASoC: TWL4030: External 256 FS clock management
Introduce a framework for a CPU or SoC DAI to request the TWL4030's
256 FS clock.  The 256 FS clock is enabled the first time it is
requested, and automatically disabled when the last DAI releases it.

A DAI may call twl4030_enable_ext_clock() multiple times to request
the 256 FS clock, but the clock is released for that DAI by calling
twl4030_disable_ext_clock() once.  If the 256 FS clock is enabled in
the DAI's hw_params callback, this prevents the clock's reference
counter from continually increasing if snd_pcm_hw_params()
is called multiple times since the device had been opened.

Change-Id: Ia873354281679c0f11a764a24f7933cf2fe5691b
Signed-off-by: Hong H. Pham <hong.pham@ti.com>