OMAP3: DSS2: DMA optimization using scaler line buffers
DISPC DMA optimization has been enabled and vrfb calls changed as required.
Optimization reduces the memory traffic (DDR memory) when rotation is set
to 90- and 270- degree and SMS-VRFB rotation engine is used.

With this change,
L3 interconnect traffic is reduced by a factor 2x for YUV422 & UYVY
DDR memory traffic is reduced by a factor 2x for YUV422 & UYVY.

This optimization is only for OMAP 36xx.

Change-Id: Ice13d7d55ae6bc0e33a89e4aa4a757dc54077347
Signed-off-by: Mikkel Christensen <mlc@ti.com>
Signed-off-by: Kishore Y <kishore.y@ti.com>
Signed-off-by: Axel Castaneda Gonzalez <x0055901@ti.com>
Signed-off-by: wonjong.lee <wonjong.lee@ti.com>