OMAP3: ISP: Force posting of pending writes in OMAP3_ISP_IOMEM_PREV ( OMAPS00239399 )
In order to force posting of pending writes, we need to write and
readback the same register, in this case the revision register.
This patch fixes a case when greengamma_table[0] is overwriten by
redgamma_table[1023], due to pending writes, which causes very dark
regions to look green on preview.

Change-Id: I2c65e9e0f8cd2bdb749f80ac1dd236b6561514db
Signed-off-by: Julian Shandorov <jshandorov@mm-sol.com>