OMAP3: GFX_FIFO_UNDERFLOW during DVFS on L3
If a rate change is performed at the same time as
the DSS accesses the SDRAM a GFX_FIFO_UNDERFLOW
is seen. To avoid this a modified patch from
K32 where the SDRAM stall waits on VSYNC is used.

Instead of looking at DSS_ICLK like in K32 this patch
looks at DSS_FCLK to tell if the DISPC registers
can be accessed.

Change-Id: I412b20e4f05a258c9cece876fb74498a9c188047
Signed-off-by: mbeckius <mbeckius@windriver.com>
Signed-off-by: Steve Schefter <sschefter@ti.com>