ASoC: Zoom2: Manage external clock for FM DAI
A dummy clock DAI is added to the TWL4030 codec to allow the Zoom2 FM
DAI to access and manage the TWL4030's external clock.  This functionality
is carried over from the 2.6.32 kernel, where all DAIs enable the
external 256FS clock on startup, and disable the clock on shutdown.
This is a work around for Errata i588.

Change-Id: I1b5d6355733b4c8d1581f34e6c45c6627aa5ed0e
Signed-off-by: Hong H. Pham <hong.pham@ti.com>