OMAP3: i2c: crash in pm_qos_remove_request()
When multiple threads are using i2c, the same handle can be added to the
QoS linked list, corrupting the list.  This patch resolves the problem by
using a stack (will be thread-specific) rather than static variable.  Since
the handle is added and removed from the list before returning to this
function, there is no need for the handle to be static.

Change-Id: I69b3b047c137fd3388dad432362aaee7b6d79e03
Signed-off-by: Steve Schefter <sschefter@ti.com>